multi: Implement Account Disable/Enable.
In this commit we implement disabling/enabling the account that is
usefull when the user wants to close all his channels and stop breez
from creating new channels automatically. This is for internal purposes.

In addition we changed the prototype of getBreezChannels in order
to retrieve both channel points and channel ids so we can reuse this
function when each is needed. This is also the cause for multiple file
changes here.