connect: retry connect to node before got channel.
In this commit we make sure the client keeps trying to connect to
routing node in case we still don' t have a channel.
Before this commit, if from some reason we got a disconect
event after the initial connection and before the channel was opened,
no further attempts were done to connect to the routing node.
This logic is required only when no  channel was opened since in case
a channel exists it is handled automatically by lightning lib.