lnnode: Refactor daemon to its own package.
This commit is a start of changing breez library code to be
more maintainable.
The attempt is to exract out code out of breez packag to separate
packages and create sub-systems that are decoupled from each other
and controlled from a central point.

In this commit the first package is lnnode that contains all the logic
to control and encapsulate (start, stop, notifications...) the lightning
deamon.