sendpayment: change API to support failed routes.
This commit changes the SendPaymentForRequest to return a structure
that contains the payment error if exists.
This result also includes the failed routes and a trace report that can
be sent for further diagnostic if the user accepts.
Due to the fact that failed routes propagate from the RPC response we
no longer need to call QueryRoutes and we use them to create the trace
report right after the SendPaymentForRequest.
This failed report is returned in the response to the client.
If the client then decides to send it for diagnositic we use the
SendPaymentFailureBugReport function which was simplified to do so.