include checkpoints difficulty adjustment headers
If an app created before block 756000 would resync from scratch, the
header sync would fail with an error calculating the difficulty
adjustment. This was because in order to calculate the expected
difficulty, the header of the previous difficulty adjustment was
inspected. This header would be empty on disk, however, so the
difficulty would be calculated with a 0 timestamp for the previous
period. This commit modifies the checkpoints, so the difficulty adjustment blocks for each checkpoint are also included.