remove record channels status
The recordChannelsStatus function would insert all zero conf channels
into the channelOpeningState bucket. This causes the funding workflow
for these channels to restart. When the funding flow restarts, LND will
start searching for the funding transaction by scanning the chain
backwards. This means as you have channels for a longer time, Breez
would consume more and more bandwidth to sync the chain from tip to
channel open height. By deleting this function, once the funding
workflow is complete once, it won't be entered again.

Upon cleaning up it seems that SetMismatchedChannels was never used. So
FetchMismatchedChannels and RemoveChannelMismatch could be removed as
well.